+++ src/frontend/apps/desk/src/features/mail-domains/aliases/components/panel/AliasesListView.tsx
@@ -1,4 +1,4 @@
-import { Button, DataGrid, SortModel } from '@openfun/cunningham-react';
+import { Button, DataGrid, SortModel } from '@gouvfr-lasuite/cunningham-react';
 import type { InfiniteData } from '@tanstack/react-query';
 import { useEffect, useMemo, useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
@@ -143,9 +143,11 @@ export function AliasesListView({
       {error && <TextErrors causes={error.cause ?? []} />}
 
       {!filteredAliases.length && (
-        <Text $align="center" $size="small" $padding={{ top: 'base' }}>
-          {t('No alias was created with this mail domain.')}
-        </Text>
+        <Box $align="center" $margin={{ top: 'base', bottom: 'base' }}>
+          <Text $size="small">
+            {t('No alias was created with this mail domain.')}
+          </Text>
+        </Box>
       )}
 
       {filteredAliases && filteredAliases.length ? (
@@ -158,7 +160,7 @@ export function AliasesListView({
                 field: 'email',
                 headerName: `${t('Alias')} • ${filteredAliases.length}`,
                 renderCell: ({ row }) => (
-                  <Text $weight="400" $theme="greyscale">
+                  <Text $weight="400" $theme="gray">
                     {row.email}
                   </Text>
                 ),
@@ -168,7 +170,7 @@ export function AliasesListView({
                 headerName: t('Destinations'),
                 enableSorting: false,
                 renderCell: ({ row }) => (
-                  <Text $weight="500" $theme="greyscale">
+                  <Text $weight="500" $theme="gray">
                     {row.count_destinations} destination
                     {row.count_destinations > 1 ? 's' : ''}
                   </Text>
@@ -189,7 +191,8 @@ export function AliasesListView({
                     <Box $direction="row" $gap="sm" $align="center">
                       {canEdit && (
                         <Button
-                          color="tertiary"
+                          color="neutral"
+                          variant="tertiary"
                           onClick={() => setEditingAliasGroup(row)}
                           style={{
                             fontWeight: '500',